Source code file renamed and photos of the product added
