+++ Source code/main.cpp
@@ -1,0 +1,9 @@
+#include <Arduino.h>
+
+void setup() {
+  // put your setup code here, to run once:
+}
+
+void loop() {
+  // put your main code here, to run repeatedly:
+}
